+++ src/empty-state-card.tsx
@@ -18,7 +18,7 @@ const useStyles = createUseStyles((theme) => ({
     },
     image: {
         height: 80,
-        [theme.mq.mobile]: {
+        [theme.mq.tabletOrSmaller]: {
             height: 64,
         },
     },
@@ -63,7 +63,7 @@ const EmptyStateCard: React.FC<Props> = ({
     const {colors} = useTheme();
     const needsButtonLinkAlignment = buttonLink && !button;
     const classes = useStyles({needsButtonLinkAlignment});
-    const {isMobile} = useScreenSize();
+    const {isTabletOrSmaller} = useScreenSize();
 
     let image;
     if (imageUrl) {
@@ -72,11 +72,11 @@ const EmptyStateCard: React.FC<Props> = ({
 
     return (
         <Boxed dataAttributes={dataAttributes}>
-            <Box paddingY={isMobile ? 24 : 40} paddingX={isMobile ? 16 : 40}>
+            <Box paddingY={isTabletOrSmaller ? 24 : 40} paddingX={isTabletOrSmaller ? 16 : 40}>
                 <section className={classes.container} aria-label={ariaLabel}>
                     <Stack space={16}>
                         {image ?? icon}
-                        <Box paddingRight={isMobile ? 48 : 0}>
+                        <Box paddingRight={isTabletOrSmaller ? 48 : 0}>
                             <Stack space={8}>
                                 <Text4 regular>{title}</Text4>
                                 <Text2 regular color={colors.textSecondary}>